Improve some minor things

+++ Bakery/Taster.py
@@ -92,7 +92,7 @@ class Taster:
 
             pickle.dump(logli_buffer, open(self.folder + "/logli_buffer.p", "wb"))
 
-        self.logli = pickle.load(open(self.folder + "/logli_buffer.p", "rb")).astype(np.float64)
+        self.logli = pickle.load(open(self.folder + "/logli_buffer.p", "rb")).astype(np.float128)
         print(self.model_param_dict["dataname"]+f"max {np.max(self.logli)}, min {np.min(self.logli)}, mean {np.mean(self.logli)}, std {np.std(self.logli)}")
 
         self.logli_average_over_samples = np.log(np.exp(self.logli).mean(axis=-1))
